# Semana_4/Día_4/Libreria/static/swagger.yaml
swagger: '2.0'
info:
  description: 'Documentación de la api de libreria'
  version: '1.0'
  title: LibreriaFlask
  termsOfService: "hhtp://www.google.com" #Pa los TyC
  contact: {
    name: "Eduardo de rivero",
    email: "jcueva12380@gmail.com"
  }
  license: {
    name: "Apache 2.0",
    url: "hhtp://www.apache...."
  }
host: 127.0.0.1:5000
basePath: /
schemes:
- https
- http
consumes:
- application/json
produces:
- application/json
paths:
...
  /buscar:
    get:
      description: filtrar los librosss
      summary: /buscar
      tags:
      - Libro
      operationId: /buscar
      deprecated: false
      produces:
      - application/json
      parameters:
      - name: palabra
        in: query
        required: true
        type: string
        description: ''
      responses:
        200:
          description: ''
          headers: {}
...